IMP_(FRNT): Add Items with a dialog
