+++ inventory-core/frontend/ConnectionTest/Program.cs
@@ -12,7 +12,7 @@ namespace ConnectionTest
             try
             {
                 Console.WriteLine("Testing gRPC connection to inventory service...");
-                
+
                 // Test the address formatting logic
                 string address = "localhost:50052";
                 string formattedAddress = address;
@@ -20,19 +20,19 @@ namespace ConnectionTest
                 {
                     formattedAddress = $"http://{address}";
                 }
-                
+
                 Console.WriteLine($"Connecting to: {formattedAddress}");
-                
+
                 // Create gRPC channel
                 using var channel = GrpcChannel.ForAddress(formattedAddress);
                 var client = new InventoryService.InventoryServiceClient(channel);
 
                 // Test the connection with a simple ping call
                 var response = await client.GetInventoryStatusAsync(new GetInventoryStatusRequest());
-                
+
                 Console.WriteLine("✅ Connection successful!");
                 Console.WriteLine($"Received response with {response.Items.Count} items");
-                
+
                 foreach (var item in response.Items)
                 {
                     Console.WriteLine($"  - {item.Name}: {item.CurrentLevel}/{item.MaxCapacity} {item.UnitId}");
