+++ inventory-core/frontend/src/InventoryClient/Models/PredictionConfigViewModel.cs
@@ -84,7 +84,7 @@ public class PredictionConfigViewModel : INotifyPropertyChanged
     protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        
+
         if (propertyName == nameof(PreferredModel))
         {
             OnPropertyChanged(nameof(PreferredModelDescription));
