+++ inventory-core/frontend/src/InventoryClient/Models/PredictionModelFitnessViewModel.cs
@@ -91,7 +91,7 @@ public class PredictionModelFitnessViewModel : INotifyPropertyChanged
     protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        
+
         if (propertyName == nameof(Model))
         {
             OnPropertyChanged(nameof(ModelDescription));
