+++ inventory-core/frontend/src/InventoryClient/Views/AddItemDialog.axaml.cs
@@ -1,0 +1,15 @@
+using Avalonia.Controls;
+using InventoryClient.ViewModels;
+
+namespace InventoryClient.Views;
+
+/// <summary>
+/// Code-behind for the AddItemDialog
+/// </summary>
+public partial class AddItemDialog : UserControl
+{
+    public AddItemDialog()
+    {
+        InitializeComponent();
+    }
+}
